Changed deckList.adds to add deck."cardName", added params to favor
Also added endTurnNoDraw which ends a players turn without drawing, so it ends the players turn with player.endTurn, then it puts the card they drew at the end of their turn back. 

Before, the deck.deckList.add() did not use the proper form to add a type card to the deck. Those cards were declared in deck, so I used deck."cardName"

Added params to player to differentiate between the targeter and the victim, made them discard the favor and the favored card.